+++ index.html
@@ -7,48 +7,6 @@
     <link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
     <link rel="stylesheet" type="text/css" href="https://cdn.ons.gov.uk/sixteens/1f354a7/css/main.css">
     <title>Statistical Business Register</title>
-    <style type="text/css">
-
-  	.node {
-      cursor: pointer;
-    }
-
-    .overlay{
-        background-color:#EEE;
-    }
-
-    .node circle {
-      fill: #fff;
-      stroke: steelblue;
-      stroke-width: 1.5px;
-    }
-
-    .node text {
-      font-size:10px;
-      font-family:sans-serif;
-    }
-
-    .link {
-      fill: none;
-      stroke: #ccc;
-      stroke-width: 1.5px;
-    }
-
-    .templink {
-      fill: none;
-      stroke: red;
-      stroke-width: 3px;
-    }
-
-    .ghostCircle.show{
-        display:block;
-    }
-
-    .ghostCircle, .activeDrag .ghostCircle{
-         display: none;
-    }
-
-  </style>
   </head>
   <body>
     <div id="root"></div>
@@ -67,5 +25,5 @@
   <!-- For the tree view -->
   <script src="http://code.jquery.com/jquery-1.10.2.min.js"></script>
   <script src="http://d3js.org/d3.v3.min.js"></script>
-  <script src="dndTree.js"></script>
+  <script src="src/resources/dndTree.js"></script>
 </html>
