+++ package.json
@@ -69,6 +69,7 @@
     "react-bootstrap": "^0.30.7",
     "react-bootstrap-button-loader": "^1.0.8",
     "react-bootstrap-table": "^3.4.5",
+    "react-d3-tree": "^1.4.0",
     "react-dom": "^15.5.4",
     "react-modal-dialog": "^4.0.7",
     "react-redux": "^5.0.5",
